Seperated the const into different files and deleted the unnecessary files
