+++ CHApp/Individual/I_Home.js
@@ -1,0 +1,30 @@
+import React, { useState } from 'react';  //State is used to store and manage data within a component.
+import { StyleSheet, TextInput, Text, View, Image, TouchableOpacity,ScrollView } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+// import CheckBox from '@react-native-community/checkbox';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import FontAwesome from '@expo/vector-icons/FontAwesome';
+import { RadioButton } from 'react-native-paper';
+import Ionicons from '@expo/vector-icons/Ionicons';
+import { Entypo } from '@expo/vector-icons';
+
+import styles from '../Style.js';
+
+const I_HomeScreen = () => {
+    return(
+      <View style={styles.greyBackground}>
+        <Text>Welcome back!</Text>
+        <TouchableOpacity style={styles.button}>
+          <Text style={styles.buttonText}>Volunteer</Text>
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.button}>
+          <Text style={styles.buttonText}>Donate</Text>
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.button}>
+          <Text style={styles.buttonText}>History</Text>
+        </TouchableOpacity>
+        </View>
+    );
+  }
+
+  export default I_HomeScreen;
